requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.4
django-environ==0.9.0
django-extensions==3.2.1
djangorestframework==3.14.0
idna==3.4
kombu==5.2.4
loguru==0.6.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
pybitrix24==1.1.0
pytz==2022.7
redis==4.4.2
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.5
